+++ weblate/checks/format.py
@@ -7,17 +7,25 @@ from __future__ import annotations
 import re
 from collections import Counter, defaultdict
 from re import Pattern
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Literal
 
 from django.utils.functional import SimpleLazyObject
-from django.utils.html import format_html_join
+from django.utils.html import format_html, format_html_join
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext, gettext_lazy
 
 from weblate.checks.base import SourceCheck, TargetCheck
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
+
+    from django_stubs_ext import StrOrPromise
+
     from weblate.trans.models import Unit
+
+    from .models import Check
+
+MissingExtraDict = dict[Literal["missing", "extra"], list[str]]
 
 PYTHON_PRINTF_MATCH = re.compile(
     r"""
@@ -102,6 +110,8 @@ PASCAL_FORMAT_MATCH = re.compile(
 
 PYTHON_BRACE_MATCH = re.compile(
     r"""
+    }(})|                                 # escaped {
+    {({)|                                 # escaped }
     {(                                  # initial {
         |                               # blank for position based
         (?P<field>
@@ -241,7 +251,7 @@ def scheme_format_is_position_based(string: str):
 
 
 def python_format_is_position_based(string: str):
-    return "(" not in string and string != "%"
+    return "(" not in string and string not in {"{", "}"}
 
 
 def name_format_is_position_based(string: str) -> bool:  # noqa: FURB118
@@ -270,13 +280,15 @@ class BaseFormatCheck(TargetCheck):
     regexp: Pattern[str]
     plural_parameter_regexp: Pattern[str] | None = None
     default_disabled = True
-    normalize_remove: str | None = None
+    normalize_remove: set[str] = set()
 
     def check_target_unit(self, sources: list[str], targets: list[str], unit: Unit):
         """Check single unit, handling plurals."""
         return any(self.check_generator(sources, targets, unit))
 
-    def check_generator(self, sources: list[str], targets: list[str], unit: Unit):
+    def check_generator(
+        self, sources: list[str], targets: list[str], unit: Unit
+    ) -> Iterable[Literal[False] | MissingExtraDict]:
         # Special case languages with single plural form
         if len(sources) > 1 and len(targets) == 1:
             yield self.check_format(sources[1], targets[0], False, unit)
@@ -326,21 +338,30 @@ class BaseFormatCheck(TargetCheck):
                 sources[1], target, len(plural_examples[i + 1]) == 1, unit
             )
 
-    def format_string(self, string: str):
+    def format_string(self, string: str) -> str:
+        """Format parsed format string into human readable value."""
         return string
 
     def cleanup_string(self, text):
         return text
 
     def normalize(self, matches: list[str]) -> list[str]:
-        if self.normalize_remove is None:
+        if not self.normalize_remove:
             return matches
-        return [m for m in matches if m != self.normalize_remove]
+        return [m for m in matches if m not in self.normalize_remove]
+
+    def extract_string(self, match: re.Match) -> str:
+        return match[0]
 
     def extract_matches(self, string: str) -> list[str]:
-        return [self.cleanup_string(x[0]) for x in self.regexp.findall(string)]
+        return [
+            self.cleanup_string(self.extract_string(x))
+            for x in self.regexp.findall(string)
+        ]
 
-    def check_format(self, source: str, target: str, ignore_missing, unit: Unit):
+    def check_format(
+        self, source: str, target: str, ignore_missing: bool, unit: Unit
+    ) -> Literal[False] | MissingExtraDict:
         """Check for format strings."""
         if not target or not source:
             return False
@@ -392,7 +413,7 @@ class BaseFormatCheck(TargetCheck):
         for match in match_objects:
             yield (match.start(), match.end(), match.group())
 
-    def format_result(self, result):
+    def format_result(self, result: MissingExtraDict) -> Iterable[StrOrPromise]:
         if (
             result["missing"]
             and all(self.is_position_based(flag) for flag in result["missing"])
@@ -411,15 +432,15 @@ class BaseFormatCheck(TargetCheck):
                     self.format_string(x) for x in set(result["extra"])
                 )
 
-    def get_description(self, check_obj):
+    def get_description(self, check_obj: Check) -> StrOrPromise:
         unit = check_obj.unit
         checks = self.check_generator(
             unit.get_source_plurals(), unit.get_target_plurals(), unit
         )
-        errors = []
+        errors: list[StrOrPromise] = []
 
         # Merge plurals
-        results = defaultdict(list)
+        results: MissingExtraDict = defaultdict(list)
         for result in checks:
             if result:
                 for key, value in result.items():
@@ -460,7 +481,7 @@ class BaseFormatCheck(TargetCheck):
 class BasePrintfCheck(BaseFormatCheck):
     """Base class for printf based format checks."""
 
-    normalize_remove = "%"
+    normalize_remove = {"%"}
 
     def __init__(self) -> None:
         super().__init__()
@@ -556,7 +577,7 @@ class SchemeFormatCheck(BasePrintfCheck):
     check_id = "scheme_format"
     name = gettext_lazy("Scheme format")
     description = gettext_lazy("Scheme format string does not match source")
-    normalize_remove = "~"
+    normalize_remove = {"~"}
 
     def format_string(self, string: str) -> str:
         return f"~{string}"
@@ -570,12 +591,44 @@ class PythonBraceFormatCheck(BaseFormatCheck):
     description = gettext_lazy("Python brace format string does not match source")
     regexp = PYTHON_BRACE_MATCH
     plural_parameter_regexp = re.compile(r"\{(?:count|number|num|n)\}")
+    normalize_remove: set[str] = {"{", "}"}
+
+    def extract_string(self, match: re.Match) -> str:
+        # 0 and 1 are escaped braces and 2 is the actual match
+        return match[0] or match[1] or match[2]
 
     def is_position_based(self, string: str):
         return name_format_is_position_based(string)
 
     def format_string(self, string: str) -> str:
         return f"{{{string}}}"
+
+    def format_result(self, result: MissingExtraDict) -> Iterable[StrOrPromise]:
+        for char in ("{", "}"):
+            if char in result["extra"]:
+                result["extra"].remove(char)
+                yield format_html(
+                    gettext("Single {} encountered in the format string."),
+                    self.format_value(char),
+                )
+        yield from super().format_result(result)
+
+    def check_format(
+        self, source: str, target: str, ignore_missing: bool, unit: Unit
+    ) -> Literal[False] | MissingExtraDict:
+        result = super().check_format(source, target, ignore_missing, unit)
+
+        noformat = PYTHON_BRACE_MATCH.sub("", target)
+
+        add_extra: list[str] = [char for char in ("{", "}") if char in noformat]
+
+        if add_extra:
+            if isinstance(result, dict):
+                result["extra"].extend(add_extra)
+            else:
+                result = {"missing": [], "extra": add_extra}
+
+        return result
 
 
 class CSharpFormatCheck(BaseFormatCheck):
@@ -622,7 +675,9 @@ class JavaMessageFormatCheck(BaseFormatCheck):
 
         return super().should_skip(unit)
 
-    def check_format(self, source: str, target: str, ignore_missing, unit: Unit):
+    def check_format(
+        self, source: str, target: str, ignore_missing: bool, unit: Unit
+    ) -> Literal[False] | MissingExtraDict:
         """Check for format strings."""
         if not target or not source:
             return False
@@ -640,7 +695,7 @@ class JavaMessageFormatCheck(BaseFormatCheck):
 
         return result
 
-    def format_result(self, result):
+    def format_result(self, result: MissingExtraDict) -> Iterable[StrOrPromise]:
         if "'" in result["missing"]:
             result["missing"].remove("'")
             yield gettext("You need to pair up an apostrophe with another one.")